Refactor to use UV for project management and for packaging as an executable
