--- requirements.txt
@@ -1,4 +1,0 @@
-requests
-requests[socks]
-urllib3
-tld
