+++ src/photon/core/colors.py
