+++ src/photon/core/mirror.py
