+++ src/photon/core/regex.py
